+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - **ESLint** Add `eslint-config-react-app` as pre-installed [#1902](https://github.com/sider/runners/pull/1902)
 - Fix checking unsupported tools [#1911](https://github.com/sider/runners/pull/1911)
+- Add `Runners::Config#exclude_branch?` method [#1913](https://github.com/sider/runners/pull/1913)
 
 ## 0.40.1